fix: fixed even more leftover refactoring errors
