+++ src/user/User.ts
@@ -1,6 +1,6 @@
 import { BaseEntity, Column, Entity, OneToMany, PrimaryColumn } from "typeorm";
 
-import { Task } from "../todo";
+import { Task } from "../task";
 
 @Entity()
 export class User extends BaseEntity {
@@ -9,6 +9,6 @@ export class User extends BaseEntity {
     @Column() name!: string;
 
     @Column() email!: string;
-    
+
     @OneToMany(() => Task, task => task.author) tasks!: Task[];
 }
